add realEstate a40

+++ frontend/modules/ad/controllers/DefaultController.php
@@ -2,9 +2,11 @@
 
 namespace frontend\modules\ad\controllers;
 
+use Yii;
 use frontend\controllers\BehaviorsController;
 use common\models\AdMain;
 use common\models\AdRealEstate;
+use yii\web\MethodNotAllowedHttpException;
 
 class DefaultController extends BehaviorsController
 {
@@ -12,9 +14,13 @@ class DefaultController extends BehaviorsController
     {
         /* @var $modelAdRealEstate \common\models\AdRealEstate */
         if($id) {
-            $modelAdRealEstate = AdRealEstate::findOne($id);
-            if($modelAdRealEstate) {
-                $modelAdRealEstate->deleteObject($modelAdRealEstate);
+            if (Yii::$app->user->can('Автор', ['model' => $modelAdRealEstate->adCategory->adMain])) {
+                $modelAdRealEstate = AdRealEstate::findOne($id);
+                if($modelAdRealEstate) {
+                    $modelAdRealEstate->deleteObject($modelAdRealEstate);
+                }
+            } else {
+                throw new MethodNotAllowedHttpException(Yii::t('app', 'You are not allowed to access this page.'));
             }
         }
